raspberry/python/requirements.txt
pybluez
pyyaml
psutil